Init multipath head namespace.
Initialize multipath head namespace. The way generic namespace is related to a generic controller, similarly multipath head namespace is related to multipath head controller. It is initialized as part of a single shared namespace enumeration. We will initialize the bio list and congestion wait queue for given multipath head namespace, this queue will be used during failover to requeue any IO returned from active path for resubmission. nvme_mpath_flush_io_work is implemented to take care of condition when both interfaces is down and IO is queued up on head mulitpath devices congestion queue. Added functionality to add and delete namespace to head multipath namespace list. This logic is similar to namespace list we have per controller. On deletion of namespaces, check if this was last namespace part of multipath head device and remove the head multipath namespace.

Signed-off-by: Anish M Jhaveri <anish.jhaveri@paviliondata.com>